+++ spring-boot-example/src/main/java/io/github/manuelarte/spring/queryparameter/jpa/example/controllers/Example1Controller.java
@@ -1,8 +1,8 @@
 package io.github.manuelarte.spring.queryparameter.jpa.example.controllers;
 
 import io.github.manuelarte.spring.queryparameter.QueryParameter;
-import io.github.manuelarte.spring.queryparameter.jpa.example.model.entities.Parent;
-import io.github.manuelarte.spring.queryparameter.jpa.example.services.ParentService;
+import io.github.manuelarte.spring.queryparameter.jpa.example.model.entities.Example1;
+import io.github.manuelarte.spring.queryparameter.jpa.example.services.Example1Service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -14,21 +14,21 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/parents")
+@RequestMapping("/api/v1/example1s")
 @lombok.AllArgsConstructor
-public class ParentController {
+public class Example1Controller {
 
-  private final ParentService parentService;
+  private final Example1Service example1Service;
 
   @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<Page<Parent>> findByPage(
+  public ResponseEntity<Page<Example1>> findByPage(
       @PageableDefault final Pageable pageable,
-      @QueryParameter(entity = Parent.class) final Specification<Parent> queryCriteria) {
-    final Page<Parent> page;
+      @QueryParameter(entity = Example1.class) final Specification<Example1> queryCriteria) {
+    final Page<Example1> page;
     if (queryCriteria != null) {
-      page = parentService.findAll(queryCriteria, pageable);
+      page = example1Service.findAll(queryCriteria, pageable);
     } else {
-      page = parentService.findAll(pageable);
+      page = example1Service.findAll(pageable);
     }
     return ResponseEntity.ok(page);
   }