+++ spring-boot-example/src/main/java/io/github/manuelarte/spring/queryparameter/jpa/example/repositories/Example1Repository.java
@@ -1,9 +1,9 @@
 package io.github.manuelarte.spring.queryparameter.jpa.example.repositories;
 
-import io.github.manuelarte.spring.queryparameter.jpa.example.model.entities.Parent;
+import io.github.manuelarte.spring.queryparameter.jpa.example.model.entities.Example1;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ParentRepository extends JpaRepository<Parent, Long>, JpaSpecificationExecutor<Parent> {
+public interface Example1Repository extends JpaRepository<Example1, Long>, JpaSpecificationExecutor<Example1> {
 
 }